Add contact groups and extProp support; add property extensions; add property handling in GDataObject; add 64-bit compilation compatibility

+++ Source/Clients/Photos/GDataEXIFTags.m
@@ -14,7 +14,7 @@
 */
 
 //
-//  GDataWebContent.m
+//  GDataEXIFTag.m
 //
 
 #import "GDataEXIFTags.h"
@@ -175,8 +175,10 @@
   
   NSMutableDictionary *dict = [NSMutableDictionary dictionary];
   NSArray *tags = [self tags];
-  
-  for (int idx = [tags count] - 1; idx >= 0; idx--) { 
+
+  // Add in reverse order so the first tag in the array wins in the case of
+  // duplicates.
+  for (NSInteger idx = [tags count] - 1; idx >= 0; idx--) { 
     
     GDataEXIFTag *tag = [tags objectAtIndex:idx];
     [dict setObject:[tag stringValue] forKey:[tag name]];
